requirements.txt
opencv-python>=4.5.0
pyautogui>=0.9.50
Pillow>=9.0.0
numpy>=1.21.0
